+++ osiris.go
@@ -1,16 +1,23 @@
 package main
 
 import (
+	"bytes"
 	"fmt"
 	"log"
 	"os"
 	"path"
 	"regexp"
 	"strings"
+	"text/template"
 
 	"github.com/fatih/color"
 	"github.com/jessevdk/go-flags"
 )
+
+type releaseOptions struct {
+	Scene      bool
+	SeriesYear bool
+}
 
 type release struct {
 	Title        string
@@ -18,12 +25,7 @@ type release struct {
 	Episode      string
 	EpisodeTitle string
 	Scene        string
-}
-
-type inputMetadata struct {
-	Film  *bool
-	Year  *string
-	Title *string
+	Options      releaseOptions
 }
 
 type args struct {
@@ -31,6 +33,8 @@ type args struct {
 	Silent     bool   `short:"s" long:"silent" description:"don't print file names"`
 	NoColor    bool   `long:"no-color" description:"disables colored output"`
 	Film       bool   `short:"f" long:"film" description:"uses film output format"`
+	SeriesYear bool   `short:"Y" long:"seriesyear" description:"whether series year is output"`
+	Scene      bool   `short:"S" long:"scene" description:"whether scene info is output"`
 	Year       string `short:"y" long:"year" description:"release year override"`
 	Title      string `short:"t" long:"title" description:"release title override"`
 	Positional struct {
@@ -38,6 +42,12 @@ type args struct {
 		Filename []string `positional-arg-name:"filename" required:"true"`
 	} `positional-args:"true"`
 }
+
+var (
+	seriesTemplate = "{{ .Title }}{{if .Options.SeriesYear}} ({{ .Year }}){{end}} - {{ .Episode}}{{if ." +
+		"EpisodeTitle}} - {{ .EpisodeTitle}}{{end}}{{if .Options.Scene }} ({{ .Scene }}){{end}}"
+	filmTemplate = "{{ .Title }} ({{ .Year }}){{if .Options.Scene }} ({{ .Scene }}){{end}}"
+)
 
 func main() {
 	var args args
@@ -51,19 +61,13 @@ func main() {
 		}
 	}
 
-	inmet := inputMetadata{
-		Film:  &args.Film,
-		Year:  &args.Year,
-		Title: &args.Title,
-	}
-
 	re, err := regexp.Compile(args.Positional.Regex)
 	if err != nil {
 		log.Fatalf("Failed to parse regex: %v\n", err)
 	}
 
 	for _, filename := range args.Positional.Filename {
-		newfilename := getFilename(filename, re, &inmet)
+		newfilename := getFilename(filename, re, &args)
 		if !args.Silent {
 			printRename(&filename, &newfilename)
 		}
@@ -73,7 +77,7 @@ func main() {
 	}
 }
 
-func getFilename(filepath string, re *regexp.Regexp, inputmetadata *inputMetadata) string {
+func getFilename(filepath string, re *regexp.Regexp, args *args) string {
 	fpath := path.Dir(filepath)
 	fext := path.Ext(filepath)
 	fnext := strings.TrimSuffix(filepath, fext)
@@ -86,34 +90,44 @@ func getFilename(filepath string, re *regexp.Regexp, inputmetadata *inputMetadat
 		}
 	}
 
+	scene := regexp.MustCompile(`\D(\.)\D`).ReplaceAllStringFunc(metadata["scene"], func(s string) string {
+		return strings.ReplaceAll(s, ".", " ")
+	})
+
 	r := release{
-		Title:        strings.ReplaceAll(metadata["title"], ".", " "),
-		Year:         metadata["year"],
-		Episode:      metadata["ep"],
+		Title:        strings.TrimSpace(strings.ReplaceAll(metadata["title"], ".", " ")),
+		Year:         strings.TrimSpace(metadata["year"]),
+		Episode:      strings.TrimSpace(metadata["ep"]),
 		EpisodeTitle: strings.TrimSpace(strings.ReplaceAll(metadata["eptitle"], ".", " ")),
-		Scene:        strings.TrimSpace(strings.ReplaceAll(metadata["scene"], ".", " ")),
+		Scene:        strings.TrimSpace(scene),
+		Options: releaseOptions{
+			Scene:      args.Scene,
+			SeriesYear: args.SeriesYear,
+		},
 	}
-	if r.Year == "" && *inputmetadata.Year != "" {
-		r.Year = *inputmetadata.Year
+	if r.Year == "" && args.Year != "" {
+		r.Year = args.Year
 	}
-	if r.Title == "" && *inputmetadata.Title != "" {
-		r.Title = *inputmetadata.Title
+	if r.Title == "" && args.Title != "" {
+		r.Title = args.Title
 	}
 
-	var newname string
-	if !*inputmetadata.Film {
-		newname = fmt.Sprintf("%s - %s", r.Title, r.Episode)
-		if r.EpisodeTitle != "" {
-			newname = fmt.Sprintf("%s - %s", newname, r.EpisodeTitle)
-		}
+	var tmpl *string
+
+	if args.Film {
+		tmpl = &filmTemplate
 	} else {
-		newname = fmt.Sprintf("%s (%s)", r.Title, r.Year)
-	}
-	if r.Scene != "" {
-		newname = fmt.Sprintf("%s (%s)", newname, r.Scene)
+		tmpl = &seriesTemplate
 	}
 
-	return fmt.Sprintf("%s/%s%s", fpath, newname, fext)
+	releaseTemplate, _ := template.New("release").Parse(*tmpl)
+	var newname bytes.Buffer
+	err := releaseTemplate.Execute(&newname, r)
+	if err != nil {
+		log.Fatalln(err)
+	}
+
+	return fmt.Sprintf("%s/%s%s", fpath, newname.String(), fext)
 }
 
 func printRename(filepath, newfilepath *string) {