+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.2.0] - 2022-06-04
+### Added
+- flags for enabling series year and scene info
+
+### Changed
+- moved to text/template
+
 ## [0.1.0] - 2022-01-21
 ### Added
 - Intitial release.